fix: dockerizando a aplicação e corrigindo problemas.
