+++ crudify-frontend-app/.eslintrc.json
@@ -1,3 +1,6 @@
 {
-  "extends": ["next/babel","next/core-web-vitals"]
+  "extends": ["next/core-web-vitals"],
+  "rules": {
+    // Você pode adicionar regras personalizadas aqui
+  }
 }
