+++ crudify-frontend-app/app/providers.tsx
@@ -1,15 +1,16 @@
 "use client"
 
 import { SessionProvider } from "next-auth/react"
-import { ThemeProvider } from "next-themes"
+import { ThemeProvider as NextThemesProvider } from "next-themes"
 import { type ThemeProviderProps } from "next-themes/dist/types"
+import React from 'react'
 
 export function Providers({ children, ...props }: ThemeProviderProps) {
   return (
     <SessionProvider>
-      <ThemeProvider {...props} enableSystem={true} attribute="class">
+      <NextThemesProvider {...props} enableSystem={true} attribute="class">
         {children}
-      </ThemeProvider>
+      </NextThemesProvider>
     </SessionProvider>
   )
 }
