+++ crudify-frontend-app/package.json
@@ -6,7 +6,7 @@
     "dev": "next dev -p 3001",
     "build": "next build",
     "start": "next start -p 3001",
-    "lint": "next lint"
+    "lint": "next lint --dir . --max-warnings=0"
   },
   "dependencies": {
     "@hookform/resolvers": "^3.9.0",
@@ -35,5 +35,17 @@
     "tailwindcss-animate": "^1.0.7",
     "typescript": "5.0.4",
     "zod": "^3.23.8"
+  },
+  "devDependencies": {
+    "@next/eslint-plugin-next": "14.2.5",
+    "@types/node": "20.2.3",
+    "@types/react": "18.2.4",
+    "@types/react-dom": "18.2.4",
+    "autoprefixer": "10.4.14",
+    "eslint": "8.41.0",
+    "eslint-config-next": "14.2.5",
+    "postcss": "8.4.41",
+    "tailwindcss": "3.3.2",
+    "typescript": "5.0.4"
   }
 }
